Revert "Fixes select word modifier for OSX" Already fixed
This reverts commit 3b0fed458046348537c8a3979251805a63f2ef85.

+++ AceJump.py
@@ -188,7 +188,7 @@ class JumpToRegionCommand(sublime_plugin.TextCommand):
 
 	def run(self, edit, start, end):
 		# Checking? try/except
-		region = sublime.Region(long(start), long(end))
+		region = sublime.Region(start, end)
 		if not region:
 			print "JumpToRegion: Bad region!"
 			return
